+++ src/main/java/gift/controller/ProductController.java
@@ -1,8 +1,7 @@
 package gift.controller;
 
-import gift.Product;
 import gift.ProductDto;
-import gift.service.ProductService;
+import gift.services.ProductService;
 import jakarta.validation.Valid;
 import java.util.List;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,7 +9,6 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
